requirements.txt
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
mysqlclient>=2.2.0
phonenumbers>=8.13.0
email-validator>=2.0.0
python-dateutil>=2.8.2 
redis~=6.2.0
email_validator~=2.2.0
numpy~=1.26.4
pillow~=10.2.0
opencv-python>=4.8.0
face-recognition>=1.3.0
whisper~=1.1.10
moviepy~=2.2.1